Revert "Fix: docker env is not available for slurm_resume"
This reverts commit d5a2ee8e5465e1de3ca1bd54d897104a47aac577.
